+++ app/html-export/page.tsx
@@ -1,0 +1,5 @@
+import HtmlExportPreview from "../../components/HtmlExportPreview";
+
+export default function HtmlExportPage() {
+  return <HtmlExportPreview />;
+}